test(fused_moe): merge GELU tests into test_moe_mlp.py
Move the GELU path helpers (_patch_npu_stream, _mock_w8a8_gelu_compute,
_GeluPathBase) and test classes (TestQuantApplyMlpGeluPath,
TestQuantApplyMlpNoGeluImpact) from the standalone test_moe_mlp_gelu.py into
test_moe_mlp.py, and delete the standalone file. All MoE MLP unit tests now
live in one module. 20 tests pass; ruff clean.
